packages/export/pyproject.toml
[project]
name = "atlcli-export"
version = "0.1.0"
description = "DOCX export for atlcli with Word template support"
requires-python = ">=3.12"
dependencies = [
    "docxtpl>=0.18.0",
    "python-docx>=1.1.0",
    "markdown>=3.5",
    "beautifulsoup4>=4.12.0",
    "docxcompose>=1.4.0",
]

[project.scripts]
atlcli-export = "atlcli_export.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atlcli_export"]

[tool.pytest.ini_options]
testpaths = ["tests"]
